move configuration into an extension
