+++ Wikipedia/Code/ArticleCollectionViewCell+WMFFeedContentDisplaying.swift
@@ -1,0 +1,44 @@
+import Foundation
+
+public extension ArticleCollectionViewCell {
+    public func configure(article: WMFArticle, contentGroup: WMFContentGroup, layoutOnly: Bool) {
+        let displayType = contentGroup.displayType()
+        
+        let imageWidthToRequest = imageView.frame.size.width < 300 ? traitCollection.wmf_nearbyThumbnailWidth : traitCollection.wmf_leadImageWidth
+        if displayType != .mainPage, let imageURL = article.imageURL(forWidth: imageWidthToRequest) {
+            isImageViewHidden = false
+            if !layoutOnly {
+                imageView.wmf_setImage(with: imageURL, detectFaces: true, onGPU: true, failure: { (error) in }, success: { })
+            }
+        } else {
+            isImageViewHidden = true
+        }
+        
+        titleLabel.text = article.displayTitle
+        isSaveButtonHidden = isSaveButtonHidden(for: displayType)
+        if displayType == .pageWithPreview {
+            descriptionLabel.text = article.wikidataDescription?.wmf_stringByCapitalizingFirstCharacter()
+            extractLabel?.text = article.snippet
+            imageViewHeight = 196
+            backgroundColor = .white
+        } else {
+            if displayType == .mainPage {
+                descriptionLabel.text = article.wikidataDescription ?? WMFLocalizedString("explore-main-page-description", value: "Main page of Wikimedia projects", comment: "Main page description that shows when the main page lacks a Wikidata description.")
+            } else {
+                descriptionLabel.text = article.wikidataDescriptionOrSnippet?.wmf_stringByCapitalizingFirstCharacter()
+            }
+            backgroundColor = backgroundColor(for: displayType)
+            extractLabel?.text = nil
+            imageViewHeight = 150
+        }
+        
+        let articleLanguage = (article.url as NSURL?)?.wmf_language
+        let articleSemanticContentAttribute = MWLanguageInfo.semanticContentAttribute(forWMFLanguage: articleLanguage)
+        titleLabel.accessibilityLanguage = articleLanguage
+        titleLabel.semanticContentAttribute = articleSemanticContentAttribute
+        descriptionLabel.accessibilityLanguage = articleLanguage
+        descriptionLabel.semanticContentAttribute = articleSemanticContentAttribute
+        extractLabel?.accessibilityLanguage = articleLanguage
+        extractLabel?.semanticContentAttribute = articleSemanticContentAttribute
+    }
+}
